# Conflicts:
#	src/main/java/com/tma/demo/entity/RoomUser.java
#	src/main/java/com/tma/demo/repository/RoomRepository.java
#	src/main/java/com/tma/demo/service/room/imp/RoomServiceImp.java

+++ src/main/java/com/tma/demo/service/comment_post/imp/FavoriteCommentServiceImp.java
@@ -2,11 +2,7 @@ package com.tma.demo.service.comment_post.imp;
 
 import com.tma.demo.common.ErrorCode;
 import com.tma.demo.common.SuccessMessage;
-import com.tma.demo.dto.request.CreateFavoriteCommentRequest;
-import com.tma.demo.dto.request.DeleteFavoriteCommentRequest;
-import com.tma.demo.dto.request.ViewListFavoriteCommentRequest;
-import com.tma.demo.dto.response.CreateFavoriteCommentResponse;
-import com.tma.demo.dto.response.ViewListFavoriteCommentResponse;
+import com.tma.demo.dto.response.LikeCommentResponse;
 import com.tma.demo.entity.Comment;
 import com.tma.demo.entity.LikeComment;
 import com.tma.demo.entity.User;
@@ -21,9 +17,8 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
-import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
@@ -35,16 +30,20 @@ public class FavoriteCommentServiceImp implements FavoriteCommentService {
     private final CommentPostService commentPostService;
 
     @Override
-    public CreateFavoriteCommentResponse createFavoriteComment(CreateFavoriteCommentRequest createFavoriteCommentRequest) {
+    public LikeCommentResponse createFavoriteComment(String commentId) {
 
         User user = userService.getUserDetails();
-        Comment comment = commentPostService.findCommentById(createFavoriteCommentRequest.getCommentId());
+        Comment comment = commentPostService.findCommentById(commentId);
+        Optional<LikeComment> existingLike = likeCommentRepository.findByUserAndComment(user, comment);
+        if (existingLike.isPresent()) {
+            throw new BaseException(ErrorCode.UNAUTHORIZED);
+        }
         LikeComment likeComment = new LikeComment();
         likeComment.setUser(user);
         likeComment.setComment(comment);
         likeComment.setCreatedAt(LocalDateTime.now());
         LikeComment savedLikeComment = likeCommentRepository.save(likeComment);
-        return new CreateFavoriteCommentResponse(
+        return new LikeCommentResponse(
                 savedLikeComment.getId().toString(),
                 user.getId().toString(),
                 comment.getId().toString(),
@@ -53,27 +52,15 @@ public class FavoriteCommentServiceImp implements FavoriteCommentService {
         );
     }
     @Override
-    public ViewListFavoriteCommentResponse getLikedCommentsByUserId(ViewListFavoriteCommentRequest request) {
-
-        Comment comment = commentPostService.findCommentById(request.getCommentId());
-        List<LikeComment> likedComments = likeCommentRepository.findByCommentId(UUID.fromString(request.getCommentId()));
-        List<ViewListFavoriteCommentResponse.LikedUserDetail> likedUserDetails = likedComments.stream()
-                .map(likeComment -> {
-                    String userId = likeComment.getUser().getId().toString();
-                    String createdAt = likeComment.getCreatedAt().toString();
-                    return new ViewListFavoriteCommentResponse.LikedUserDetail(userId, createdAt);
-                })
-                .collect(Collectors.toList());
-        return new ViewListFavoriteCommentResponse(request.getCommentId(), likedUserDetails);
-    }
-    @Override
-    public String deleteFavoriteComment(DeleteFavoriteCommentRequest deleteFavoriteCommentRequest) {
+    public String deleteFavoriteComment(String likeCommentId) {
         User user = userService.getUserDetails();
-        Comment comment = commentPostService.findCommentById(deleteFavoriteCommentRequest.getCommentId());
-        LikeComment likeComment = likeCommentRepository.findByUserAndComment(user, comment)
+        LikeComment likeComment = likeCommentRepository.findById(UUID.fromString(likeCommentId))
                 .orElseThrow(() -> new BaseException(ErrorCode.LIKE_DOES_NOT_EXIST));
+
+        if (!likeComment.getUser().getId().equals(user.getId())) {
+            throw new BaseException(ErrorCode.UNAUTHORIZED);
+        }
         likeCommentRepository.delete(likeComment);
         return SuccessMessage.DELETE_FAVOURITE_POST_SUCCESS.getMessage();
     }
-
 }
